rethink parser
Parser now uses some owned values for simplicity.
It can still avoid allocations in many cases, such as for well known
headers and parts of the request line.

Many fields like "max_line_len_bytes" are removed from the Parser type.
This is a library designed to be called by another library, not by the
person writing the actual server software, so it makes sense to allow
the parent library to handle these things and focus purely on the
protocol.